+++ bases/models.py
@@ -1,0 +1,11 @@
+from django.db import models
+
+# Create your models here.
+
+
+class BaseTimeModel(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True)
+    modified_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        abstract = True
